+++ src/Files.App/Dialogs/SettingsDialog.xaml.cs
@@ -51,8 +51,9 @@ namespace Files.App.Dialogs
 				0 => SettingsContentFrame.Navigate(typeof(Appearance)),
 				1 => SettingsContentFrame.Navigate(typeof(Preferences)),
 				2 => SettingsContentFrame.Navigate(typeof(Folders)),
-				3 => SettingsContentFrame.Navigate(typeof(Advanced)),
-				4 => SettingsContentFrame.Navigate(typeof(About)),
+				3 => SettingsContentFrame.Navigate(typeof(TagsSettings)),
+				4 => SettingsContentFrame.Navigate(typeof(Advanced)),
+				5 => SettingsContentFrame.Navigate(typeof(About)),
 				_ => SettingsContentFrame.Navigate(typeof(Appearance))
 			};
 		}
