+++ src/Files.App/ViewModels/SettingsViewModels/AdvancedViewModel.cs
@@ -7,15 +7,12 @@ using Files.App.Filesystem.StorageItems;
 using Files.App.Helpers;
 using Files.App.Shell;
 using Files.Backend.Services.Settings;
-using Files.Backend.ViewModels.FileTags;
 using Files.Shared.Extensions;
 using Microsoft.Win32;
 using SevenZip;
 using System;
-using System.Collections.ObjectModel;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Input;
@@ -31,25 +28,14 @@ namespace Files.App.ViewModels.SettingsViewModels
 		private IUserSettingsService UserSettingsService { get; } = Ioc.Default.GetRequiredService<IUserSettingsService>();
 
 		private readonly IBundlesSettingsService bundlesSettingsService = Ioc.Default.GetRequiredService<IBundlesSettingsService>();
-
+		
 		private readonly IFileTagsSettingsService fileTagsSettingsService = Ioc.Default.GetRequiredService<IFileTagsSettingsService>();
-
-		private bool isBulkOperation = true;
-
-		private bool isDragStarting = true;
 
 		public ICommand SetAsDefaultExplorerCommand { get; }
 		public ICommand SetAsOpenFileDialogCommand { get; }
 		public ICommand ExportSettingsCommand { get; }
 		public ICommand ImportSettingsCommand { get; }
 		public ICommand OpenSettingsJsonCommand { get; }
-		public ICommand AddTagCommand { get; }
-		public ICommand SaveNewTagCommand { get; }
-		public ICommand CancelNewTagCommand { get; }
-
-		public NewTagViewModel NewTag = new();
-
-		public ObservableCollection<ListedTagViewModel> Tags { get; set; }
 
 		public AdvancedViewModel()
 		{
@@ -61,35 +47,6 @@ namespace Files.App.ViewModels.SettingsViewModels
 			ExportSettingsCommand = new AsyncRelayCommand(ExportSettings);
 			ImportSettingsCommand = new AsyncRelayCommand(ImportSettings);
 			OpenSettingsJsonCommand = new AsyncRelayCommand(OpenSettingsJson);
-
-			// Tags Commands
-			AddTagCommand = new RelayCommand(DoAddNewTag);
-			SaveNewTagCommand = new RelayCommand(DoSaveNewTag);
-			CancelNewTagCommand = new RelayCommand(DoCancelNewTag);
-
-			Tags = new ObservableCollection<ListedTagViewModel>();
-			Tags.CollectionChanged += Tags_CollectionChanged;
-			fileTagsSettingsService.FileTagList?.ForEach(tag => Tags.Add(new ListedTagViewModel(tag)));
-
-			isBulkOperation = false;
-		}
-
-		private void Tags_CollectionChanged(object? sender, System.Collections.Specialized.NotifyCollectionChangedEventArgs e)
-		{
-			if (isBulkOperation)
-				return;
-			
-			// Reaordering ListView has no events, but its collection is updated twice,
-			// first to remove the selected item, and second to add the item at the selected position.
-			if (isDragStarting)
-			{
-				isDragStarting = false;
-				return;
-			}
-
-			isDragStarting = true;
-
-			fileTagsSettingsService.FileTagList = Tags.Select(tagVM => tagVM.Tag).ToList();
 		}
 
 		private async Task OpenSettingsJson()
@@ -322,13 +279,6 @@ namespace Files.App.ViewModels.SettingsViewModels
 			set => SetProperty(ref isSetAsOpenFileDialog, value);
 		}
 
-		private bool isCreatingNewTag;
-		public bool IsCreatingNewTag
-		{
-			get => isCreatingNewTag;
-			set => SetProperty(ref isCreatingNewTag, value);
-		}
-
 		private FileSavePicker InitializeWithWindow(FileSavePicker obj)
 		{
 			WinRT.Interop.InitializeWithWindow.Initialize(obj, App.WindowHandle);
@@ -341,82 +291,6 @@ namespace Files.App.ViewModels.SettingsViewModels
 			WinRT.Interop.InitializeWithWindow.Initialize(obj, App.WindowHandle);
 
 			return obj;
-		}
-
-		private void DoAddNewTag()
-		{
-			NewTag.Reset();
-			IsCreatingNewTag = true;
-		}
-
-		private void DoSaveNewTag()
-		{
-			IsCreatingNewTag = false;
-
-			fileTagsSettingsService.CreateNewTag(NewTag.Name, NewTag.Color);
-
-			isBulkOperation = true;
-			Tags.Clear();
-			fileTagsSettingsService.FileTagList?.ForEach(tag => Tags.Add(new ListedTagViewModel(tag)));
-			isBulkOperation = false;
-		}
-
-		private void DoCancelNewTag()
-		{
-			IsCreatingNewTag = false;
-		}
-
-		public void EditExistingTag(ListedTagViewModel item, string newName, string color)
-		{
-			fileTagsSettingsService.EditTag(item.Tag.Uid, newName, color);
-
-			isBulkOperation = true;
-			Tags.Clear();
-			fileTagsSettingsService.FileTagList?.ForEach(tag => Tags.Add(new ListedTagViewModel(tag)));
-			isBulkOperation = false;
-		}
-
-		public void DeleteExistingTag(ListedTagViewModel item)
-		{
-			isBulkOperation = true;
-			Tags.Remove(item);
-			isBulkOperation = false;
-
-			fileTagsSettingsService.DeleteTag(item.Tag.Uid);
-		}
-	}
-
-	public class NewTagViewModel : ObservableObject
-	{
-		private string name = string.Empty;
-		public string Name
-		{
-			get => name;
-			set
-			{
-				if (SetProperty(ref name, value))
-					OnPropertyChanged(nameof(IsNameValid));
-			}
-		}
-
-		private string color = "#FFFFFFFF";
-		public string Color
-		{
-			get => color;
-			set => SetProperty(ref color, value);
-		}
-
-		private bool isNameValid;
-		public bool IsNameValid
-		{
-			get => isNameValid;
-			set => SetProperty(ref isNameValid, value);
-		}
-
-		public void Reset()
-		{
-			Name = string.Empty;
-			Color = ColorHelpers.RandomColor();
 		}
 	}
 }
